refactor(project): naming 리펙토링
