+++ src/View/getUserInput.js
@@ -1,10 +1,10 @@
 import { Console } from '@woowacourse/mission-utils';
-import { EMPTY_INPUT_MESSAGE } from '../constant/errorMessage';
+import { TYPE_INPUT_MESSAGE } from '../constant/errorMessage.js';
 
 export const getUserInput = (message) => {
   const userInput = Console.readLineAsync(message + '\n');
   if (!userInput) {
-    return new Error(`[Error]: ${EMPTY_INPUT_MESSAGE}}`);
+    throw new Error(`[ERROR]: ${TYPE_INPUT_MESSAGE}}`);
   }
   return userInput;
 };
