+++ src/View/printDistance.js
@@ -5,3 +5,4 @@ const ONE_STEP = '-';
 export const printDistance = (car) => {
   Console.print(car.name + ' : ' + ONE_STEP.repeat(car.distance));
 };
+
