+++ src/View/printWinner.js
@@ -1,5 +1,5 @@
 import { Console } from '@woowacourse/mission-utils';
-import { FINER_WINNER } from '../constant/message.js';
+import { FINER_WINNER } from '../constant/raceMessage.js';
 
 export const printWinner = (nameArray) => {
   Console.print(
