+++ src/constant/errorMessage.js
@@ -1,0 +1,5 @@
+export const TYPE_INPUT_MESSAGE = '값을 입력해주세요'
+
+export const TYPE_POSITIVE_INTEGER = '양의 정수를 입력해주세요'
+
+export const TYPE_CORRECT_LENGTH = '5글자 이하의 이름을 입력해주세요'
